pyproject.toml
[project]
name = "mujocollm"
version = "0.1.0"
description = "Language-guided task abstraction for pick-and-place in MuJoCo with Franka Panda."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mujoco>=3.4.0",
    "numpy>=1.24.0",
    "pydantic>=2.7.0",
    "google-genai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
mujoco-env-smoke = "mujoco_env.env:main"
segment-dataset = "segmentation.segment:main"
generate-task-graph = "llm_abstraction.generate_task_graph:main"

[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "mujoco_env",
    "segmentation",
    "llm_abstraction",
    "utils",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
